backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
langchain==0.1.0
langchain-groq==0.1.0
langchain-community==0.0.10
pymongo==4.6.0
PyMySQL==1.1.0
SQLAlchemy==2.0.23
pandas==2.1.3
python-multipart==0.0.6
httpx==0.25.2 